Move CLI args to dedicated module
